# perf(remix-server-runtime): use json.parse for route manifest