--- a/.changeset/two-pumpkins-repair.md
+++ b/.changeset/two-pumpkins-repair.md
@@ -1,0 +1,5 @@
+---
+"@remix-run/server-runtime": patch
+---
+
+Improve performance by using JSON.parse instead of an object literal loading the route manifest on the client.