@pcattori 2023-07-05
`JSON.parse` vs `JSON.stringify` is not a bottleneck and only makes sub-millisecond improvement, but adds additional indirection. Feel free to reopen if this becomes a bottleneck.