@brophdawg11 2023-07-05
Just as a note, I don't think we can fully trust the benchmarks reported here since they're from Node on a 16-core machine.  The thing aiming to be sped up here is the cost of reading the manifest in the browser on the client device, not necessarily a 16-core node CPU.  I would expect to see better results on a slower device/CPU, but would want to see real RUM data before declaring this approach as "better".